fix for octal needed to use unsigned int

+++ number.c
@@ -81,7 +81,7 @@ int print_o(char **buffer, int *total_chars, va_list list)
 	int i;
 	char *word;
 
-	word = convert(va_arg(list, long  int), 8);
+	word = convert(va_arg(list, unsigned int), 8);
 
 	for (i = 0; word[i] != '\0'; ++i)
 	{
